requirements.txt
polars
youtube-transcript-api
google-api-python-client
groq
discordwebhook
python-dotenv
google-genai
spacy
langgraph
pydantic
pydantic-settings
tenacity
python-json-logger
tiktoken
sqlmodel
httpx
pytest
prometheus_client